refactor: move `insertFinalNewLine` command from `common` folder
